third_party/WebKit/LayoutTests/fast/scrolling/scrolling-apis-subpixel.html

<!DOCTYPE html>
<style>
.spacer {
  height: 1000px;
  width: 1000px;
}
#scroller, body {
  height: 100px;
  width: 100px;
  overflow: scroll;
}

</style>
<body class=scroller>
<div id=scroller>
    <div class=spacer></div>
</div>
<div class=spacer></div>
</body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
// Note we current support fractional scrolling only for the special case of
// browser zoom.  When http://crbug.com/414283 is fixed, we should test
// other cases like device scale.

// FIXME: Make this smaller. crbug.com/414283.
var floatPrecision = 0.01;

function testScroll(scrollOffset, expectedScrollOffset) {
  // Scrolling DIV with scrollTop/scrollLeft
  scroller.scrollTop = scrollOffset;
  assert_approx_equals(scroller.scrollTop, expectedScrollOffset, floatPrecision);
  scroller.scrollLeft = scrollOffset;
  assert_approx_equals(scroller.scrollLeft, expectedScrollOffset, floatPrecision);

  // Note that the body element is a special case - we don't attempt to
  // test it here as it's semantics are in flux (http://goo.gl/BFHtMR).

  //Scrolling the document with window.scroll
  window.scroll(0,0);
  scrollOffset++;
  expectedScrollOffset++;
  window.scroll(scrollOffset, scrollOffset);
  assert_approx_equals(window.scrollY, expectedScrollOffset, floatPrecision);
  assert_approx_equals(window.scrollX, expectedScrollOffset, floatPrecision);

  // Scrolling the document with window.scrollTo
  window.scroll(0,0);
  window.scrollTo(scrollOffset, scrollOffset);
  assert_approx_equals(window.pageYOffset, expectedScrollOffset, floatPrecision);
  assert_approx_equals(window.pageXOffset, expectedScrollOffset, floatPrecision);

  // Scrolling the document with window.scrollBy
  window.scroll(1,1);
  window.scrollBy(scrollOffset - 1, scrollOffset - 1);
  assert_approx_equals(window.scrollY, expectedScrollOffset, floatPrecision);
  assert_approx_equals(window.scrollX, expectedScrollOffset, floatPrecision);
}

function testPageZoom(zoom) {
  internals.setZoomFactor(2);
  testScroll(4, 4);
  testScroll(4.5, 4.5);
}

test( () => {
  internals.settings.setPreferCompositingToLCDTextEnabled(false);
  testScroll(4.2, 4);
  testPageZoom(2);
}, "By default we don't yet get fractional scroll offsets with more " +
   "granularity than the page zoom factor.");

if (internals.runtimeFlags.fractionalScrollOffsetsEnabled) {
  test( () => {
    internals.settings.setPreferCompositingToLCDTextEnabled(true);
    testScroll(4.2, 4.2);
    testPageZoom(2);
  }, "Verifies that scrolling APIs support fractional offsets.");
}
</script>
